requirements.txt
asgiref==3.2.7
beautifulsoup4==4.6.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
djangotoolbox==1.8.0
future==0.16.0
idna==2.5
ipython-genutils==0.2.0
jsondiff==1.1.1
nose==1.3.7
pbr==5.4.5
py==1.9.0
PyJWT==1.7.1
pytest==3.2.1
python-dateutil==2.6.1
pytz==2019.3
requests==2.18.1
six==1.10.0
SQLAlchemy==1.1.12
sqlparse==0.3.1
traitlets==4.3.2
unittest-tap-reporting==0.2
unittest-xml-reporting==2.1.0
urllib3==1.21.1
virtualenv==15.1.0
virtualenv-clone==0.2.6

